+++ packages/core/src/telemetry/telemetry.ts
@@ -2,12 +2,12 @@ import {
   context as otlpContext,
   SpanStatusCode,
   trace,
-  Tracer,
+  type Tracer,
   propagation,
-  SpanOptions,
-  Context,
+  type SpanOptions,
+  type Context,
   context,
-  Span,
+  type Span,
 } from '@opentelemetry/api';
 import { getNodeAutoInstrumentations } from '@opentelemetry/auto-instrumentations-node';
 import { OTLPTraceExporter as OTLPHttpExporter } from '@opentelemetry/exporter-trace-otlp-http';
@@ -19,11 +19,11 @@ import {
   TraceIdRatioBasedSampler,
   AlwaysOnSampler,
   AlwaysOffSampler,
-  Sampler,
+  type Sampler,
 } from '@opentelemetry/sdk-trace-base';
 import { ATTR_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
 
-import { OtelConfig } from './types';
+import { type OtelConfig } from './types';
 import { hasActiveTelemetry } from './utility';
 
 export { OTLPTraceExporter as OTLPStorageExporter } from './storage-exporter';
@@ -71,6 +71,7 @@ export class Telemetry {
     this.name = config.serviceName ?? 'default-service';
 
     // Only initialize in server environment
+    // @ts-ignore window is not defined in node
     if (typeof window === 'undefined') {
       // In development, always create a new instance
       // In production, use existing instance if available