+++ packages/evals/src/attachListeners.ts
@@ -5,16 +5,25 @@ import { MastraStorage } from '@mastra/core/storage';
 import { GLOBAL_RUN_ID_ENV_KEY } from './constants';
 
 export async function attachListeners(mastra?: Mastra) {
+  if (mastra?.storage) {
+    await mastra.storage.init();
+  }
+
   registerHook(AvailableHooks.ON_EVALUATION, async traceObject => {
-    if (mastra?.memory?.storage) {
-      await mastra.memory.storage.insert({
+    if (mastra?.storage) {
+      await mastra.storage.insert({
         tableName: MastraStorage.TABLE_EVALS,
         record: {
-          result: JSON.stringify(traceObject.result),
-          meta: JSON.stringify(traceObject.meta),
           input: traceObject.input,
           output: traceObject.output,
-          createdAt: new Date().toISOString(),
+          result: JSON.stringify(traceObject.result),
+          agent_name: traceObject.agentName,
+          metric_name: traceObject.metricName,
+          instructions: traceObject.instructions,
+          test_info: traceObject.testInfo ? JSON.stringify(traceObject.testInfo) : null,
+          global_run_id: traceObject.globalRunId,
+          run_id: traceObject.runId,
+          created_at: new Date().toISOString(),
         },
       });
     }